pkg/manifests: move cluster_k8s_io to manifests from machines.
This is to break an import loop between pkg/assets/machines and
pkg/assets/manifests.
